% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemness.score.R
\name{stemness.score}
\alias{stemness.score}
\title{Stemness Score of Glioma Cells Using scATAC-seq Data}
\usage{
stemness.score(
  main_folder_name,
  matrix_file_name,
  peaks_file_name,
  barcodes_file_name,
  gene_list_file_name,
  importance_file_name,
  genomic_regions
)
}
\arguments{
\item{main_folder_name}{the name of the folder where three files for a patient stored}

\item{matrix_file_name}{matrix file name in MEX (Market Exchange Format) format}

\item{peaks_file_name}{BED file name containing peaks}

\item{barcodes_file_name}{tsv file name containing barcodes}

\item{gene_list_file_name}{csv file name containing gene list}

\item{importance}{a data frame composed of genes and their importance}

\item{genomic.regions}{a data frame obtained from GeneAlaCart}
}
\value{
a data frame
}
\description{
The \code{stemness.score} function estimates the stemness index of
each glioma cell using scATAC-seq data.
}
\author{
Necla Kochan and Gokhan Karakulah
}
\keyword{glioblastoma}
\keyword{stemness}
